Refactor: 움직임 탐지관련 기능 모듈화 진행
