+++ services/detectMovement.js
@@ -1,36 +1,97 @@
+/* eslint-disable no-await-in-loop */
+const path = require("path");
 const sharp = require("sharp");
 
-const BLACK_THRESHOLD = 11;
-const MOVEMENT_THRESHOLD = 0.09;
+const ensureDir = require("../util/ensureDir");
+const analyzeSingleMovement = require("./analyzeSingleMovement");
 
-async function detectMovement(imagePath) {
-  const { data, info } = await sharp(imagePath)
-    .raw()
-    .toBuffer({ resolveWithObject: true });
+const LOW_FRAME_WIDTH = 250;
+const LOW_FRAME_HEIGHT = 250;
+const BLACK_THRESHOLD = 20;
 
-  let blackPixelCount = 0;
+async function detectMovement(filteredFramePathList, folderName) {
+  const outputFileDirectory = path.join(
+    __dirname,
+    `../temp/difference/${folderName}`,
+  );
 
-  for (let i = 0; i < data.length; i += info.channels) {
-    const red = data[i];
-    const green = data[i + 1];
-    const blue = data[i + 2];
+  await ensureDir(outputFileDirectory);
 
-    if (
-      red < BLACK_THRESHOLD &&
-      green < BLACK_THRESHOLD &&
-      blue < BLACK_THRESHOLD
-    ) {
-      blackPixelCount += 1;
+  const detectedMovementRatioList = [];
+
+  for (
+    let indexOf1fps = 0;
+    indexOf1fps < filteredFramePathList.length - 1;
+    indexOf1fps += 1
+  ) {
+    const currentFramePath = filteredFramePathList[indexOf1fps];
+    const nextFramePath = filteredFramePathList[indexOf1fps + 1];
+    const frameNumber = parseInt(
+      path.basename(currentFramePath).split("_").pop(),
+      10,
+    );
+
+    const outputFileName = `difference_${frameNumber}.jpg`;
+    const outputPath = path.join(outputFileDirectory, outputFileName);
+
+    const currentLowFrame = sharp(currentFramePath)
+      .resize(LOW_FRAME_WIDTH, LOW_FRAME_HEIGHT)
+      .modulate({ saturation: 0.1 })
+      .modulate({ brightness: 1.9 });
+
+    const nextLowFrame = sharp(nextFramePath)
+      .resize(LOW_FRAME_WIDTH, LOW_FRAME_HEIGHT)
+      .modulate({ saturation: 0.1 })
+      .modulate({ brightness: 1.9 });
+
+    const resultOfDifference = await currentLowFrame
+      .composite([
+        { input: await nextLowFrame.toBuffer(), blend: "difference" },
+      ])
+      .modulate({ saturation: 3.0 })
+      .toColourspace("b-w")
+      .toBuffer();
+
+    const { data, info } = await sharp(resultOfDifference)
+      .raw()
+      .toBuffer({ resolveWithObject: true });
+
+    for (let i = 0; i < data.length; i += info.channels) {
+      let red = data[i];
+      let green = data[i + 1];
+      let blue = data[i + 2];
+
+      const isDark =
+        red < BLACK_THRESHOLD &&
+        green < BLACK_THRESHOLD &&
+        blue < BLACK_THRESHOLD;
+      const colorValue = isDark ? 0 : 255;
+
+      red = colorValue;
+      green = colorValue;
+      blue = colorValue;
     }
+
+    await sharp(data, {
+      raw: {
+        width: info.width,
+        height: info.height,
+        channels: info.channels,
+      },
+    })
+      .resize(400, 400)
+      .extract({ width: 200, height: 200, left: 100, top: 100 })
+      .toFile(outputPath);
+
+    const resultOfMovementRatio = await analyzeSingleMovement(outputPath);
+
+    detectedMovementRatioList.push({
+      path: currentFramePath,
+      movementRatio: resultOfMovementRatio,
+    });
   }
 
-  const blackRatio = blackPixelCount / (info.width * info.height);
-
-  if (blackRatio > MOVEMENT_THRESHOLD) {
-    return true;
-  }
-
-  return false;
+  return detectedMovementRatioList;
 }
 
 module.exports = detectMovement;
